requirements.txt
discord.py
aiosqlite
python-dotenv
Flask
aiohttp